+++ src/modules/getConfig/index.ts
@@ -1,7 +1,7 @@
 import { Prefix } from "../../shared/const/Prefix"
 import { getConfigRules } from "../getConfigRules"
 
-import type { MasterRecord, Options, RuleName } from "../../shared/types"
+import type { RuleName } from "../../shared/types/RuleName"
 
 type GetConfigs = (master: MasterRecord[]) => {
   plugins: string[]
