+++ src/modules/getConfigRules/index.ts
@@ -1,6 +1,6 @@
 import { Prefix } from "../../shared/const/Prefix"
 
-import type { UtilFunction, Options, RuleName } from "../../shared/types"
+import type { RuleName } from "../../shared/types/RuleName"
 
 type GetConfigRules = UtilFunction<Record<RuleName, Options>>
 export const getConfigRules: GetConfigRules = (current, master) => {
