+++ src/modules/getRules/index.ts
@@ -1,4 +1,4 @@
-import type { UtilFunction, EsLintRuleBase, RuleName } from "../../shared/types"
+import type { RuleName } from "../../shared/types/RuleName"
 
 type GetRules = UtilFunction<Record<RuleName, EsLintRuleBase>>
 export const getRules: GetRules = (current, master) => {
