+++ src/shared/types/RuleName/index.ts
@@ -1,0 +1,1 @@
+export type RuleName = string
