+++ src/shared/types/index.ts
@@ -1,6 +1,6 @@
+import type { RuleName } from "./RuleName"
 import type { TSESLint } from "@typescript-eslint/utils"
 
-export type RuleName = string
 export type Options = readonly unknown[]
 type MessageId = string
 export type EsLintRuleBase = TSESLint.RuleModule<MessageId, Options>
